pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dbt-bigquery>=1.11.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.1.1",
    "pandas>=3.0.1",
    "prefect>=3.6.19",
    "requests>=2.32.5",
]
